refactor : refactor tcp, ssh
