+++ java-module/mc-o11y-manager/src/test/java/com/innogrid/tabcloudit/o11ymanager/service/TcpServiceTest.java
@@ -5,7 +5,6 @@ import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import com.mcmp.o11ymanager.model.host.HostStatus;
-import com.mcmp.o11ymanager.oldService.domain.TcpServiceImpl;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.DisplayName;
